hasselblad_load_raw: fix warning

+++ rtengine/dcraw.cc
@@ -2441,8 +2441,8 @@ void CLASS hasselblad_load_raw()
         for (int col = 0; col < raw_width; col += 2) {
             for (int s = 0; s < tiff_samples * 2; s += 2) {
                 const int len[2]= {
-                    ph1_huff(jh.huff[0]),
-                    ph1_huff(jh.huff[0])
+                    static_cast<int>(ph1_huff(jh.huff[0])),
+                    static_cast<int>(ph1_huff(jh.huff[0]))
                 };
                 for (int c = 0; c < 2; ++c) {
                     diff[s + c] = hb_bits(len[c]);
